feat: move integration test to root; move event-enums to event module
